+++ src/main/java/Controller/ControllerOperatore.java
@@ -13,7 +13,7 @@ public class ControllerOperatore {
     public static final int MOVIMENTO_REGISTRATO_SOTTO_SCORTA=2;
     public static final int ERRORE_QUANTITA_INSUFFICIENTE=-1;
 
-    public static int registraMovimento(String tipo, int quantita, Long id_prodotto, int id_Operatore){
+    public static int registraMovimento(String tipo, int quantita, int id_prodotto, int id_Operatore){
 
         RegistroMovimenti registroMovimenti = new RegistroMovimenti();
         RegistroProdotti registroProdotti = new RegistroProdotti();
